+++ HISTORY.rst
@@ -235,6 +235,7 @@
 - Generalize the document sandbox to work with other models.
 - Enable template sandbox for document files and document versions.
 - Add template sandbox API view.
+- Add model property information tool view.
 
 4.8.3 (2024-10-04)
 ==================